Move `set_pullupdown` to `Pin`.
